correctly expand __FILE__ on a server-side by using relative paths
Before, a client send cppInFile as an abs file name, and a server launched
> g++ /{clientWorkingDir}/{cppInFile}
Now, a client sends cppInFile exactly the same as it was specified in cmd line,
and besides, it sends cwd.
On a server, we launch g++ in a dir {clientWorkingDir}/{cwd} and pass
> g++ {cppInFile}
directly, as a relative one, which leads to __FILE__
expanded equally to client launch.